requirements.txt
# SophisticadoTools Pip Requirements
python-dotenv
requests
requests_cache
web3
